+++ website/app/themes/2bdm/assets/scripts/behavior_filter_articles.js
@@ -4,53 +4,40 @@ const behavior_filter_articles = {
         const loadMoreButton = document.getElementById('load-more');
         const articlesContainer = document.querySelector('.next-articles-container');
 
-        let selectedTerms = [];
+        let selectedTerm = 'all';
         let currentPage = 1;
 
         terms.forEach(term => {
             term.addEventListener('click', function() {
                 const termId = this.getAttribute('data-term-id');
 
-                if (termId === 'all') {
-                    selectedTerms = [];
-                    terms.forEach(t => t.classList.remove('selected'));
-                    this.classList.add('selected');
-                } else {
-                    if (selectedTerms.includes(termId)) {
-                        selectedTerms = selectedTerms.filter(id => id !== termId);
-                        this.classList.remove('selected');
-                    } else {
-                        selectedTerms.push(termId);
-                        this.classList.add('selected');
-                    }
-                    document.querySelector('.filter-term[data-term-id="all"]').classList.remove('selected');
-                }
+                // reset the filter visually
+                terms.forEach(t => t.classList.remove('selected'));
+                this.classList.add('selected');
 
-                currentPage = 1; // Reset to first page on new filter
-                fetchArticles(selectedTerms, currentPage);
+                // We can select only one filter
+                selectedTerm = termId;
+
+                currentPage = 1;
+                fetchArticles(selectedTerm, currentPage);
             });
         });
 
         if (loadMoreButton) {
             loadMoreButton.addEventListener('click', function() {
                 currentPage++;
-                fetchArticles(selectedTerms, currentPage);
+                fetchArticles(selectedTerm, currentPage);
             });
         }
 
-        function fetchArticles(terms, page) {
+        function fetchArticles(term, page) {
             const ajaxUrl = loadMoreButton.getAttribute('data-url');
+
             let body = {
                 action: 'load_more_or_filtered_articles',
                 paged: page,
-            }
-
-            // We send the terms only if there are in the request
-            if (terms.length) {
-                body.terms = terms.join(',');
-            } else {
-                body.terms = 'all';
-            }
+                terms: term // the key name is plural in case we need to go back to multiple selection filter
+            };
 
             fetch(ajaxUrl, {
                 method: 'POST',
@@ -62,13 +49,14 @@ const behavior_filter_articles = {
                 .then(response => response.json())
                 .then(jsonResponse => {
                     const data = jsonResponse.data;
+
                     if (page === 1) {
                         articlesContainer.innerHTML = data.articles_html;
                     } else {
                         articlesContainer.insertAdjacentHTML('beforeend', data.articles_html);
                     }
 
-                    if (data.remaining_articles < 0) {
+                    if (data.remaining_articles <= 0) {
                         loadMoreButton.style.display = 'none';
                     } else {
                         loadMoreButton.style.display = 'inline-block';